requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-Security-Too==5.3.2
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
pyotp==2.9.0
Pillow>=9.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0.post1
python-magic==0.4.27
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
SQLAlchemy>=2.0.27
qrcode>=7.4.2
